Move diagnostics tests to their own file
